requirements.txt
emoji
feedparser
future
psycopg2-binary
python-telegram-bot==11.1.0
requests
sqlalchemy
